# server/src/includes/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Hakanai API",
    "description": "Zero-knowledge one-time secret sharing service",
    "version": "1.7.0",
    "contact": {
      "url": "https://github.com/czerwonk/hakanai"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/czerwonk/hakanai/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "{protocol}://{host}",
      "description": "Hakanai server",
      "variables": {
        "protocol": {
          "enum": ["http", "https"],
          "default": "https"
        },
        "host": {
          "default": "hakanai.example.com"
        }
      }
    }
  ],
  "paths": {
    "/api/v1/secret": {
      "post": {
        "summary": "Create a new secret",
        "description": "Creates a new one-time secret with optional expiration. All encryption happens client-side.",
        "operationId": "createSecret",
        "security": [
          {},
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSecretRequest"
              },
              "example": {
                "data": "bXkgc2VjcmV0IG1lc3NhZ2U=",
                "expires_in": 3600
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Secret created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateSecretResponse"
                },
                "example": {
                  "id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid JSON or missing required fields",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid authentication token",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/secret/{id}": {
      "get": {
        "summary": "Retrieve a secret",
        "description": "Retrieves a one-time secret by ID. The secret is deleted after retrieval.",
        "operationId": "getSecret",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "UUID of the secret to retrieve",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "responses": {
          "200": {
            "description": "Secret retrieved successfully",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "description": "Base64-encoded secret data"
                },
                "example": "bXkgc2VjcmV0IG1lc3NhZ2U="
              }
            }
          },
          "404": {
            "description": "Secret not found or has expired",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "410": {
            "description": "Secret has already been accessed and deleted",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/tokens": {
      "post": {
        "summary": "Create a new user token",
        "description": "Creates a new user token with optional size limit and TTL. Requires admin authentication.",
        "operationId": "createUserToken",
        "security": [
          {
            "adminAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTokenRequest"
              },
              "example": {
                "upload_size_limit": 1048576,
                "ttl_seconds": 2592000
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTokenResponse"
                },
                "example": {
                  "token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid JSON or missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid admin token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Invalid admin token"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error - failed to create token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateSecretRequest": {
        "type": "object",
        "required": ["data", "expires_in"],
        "properties": {
          "data": {
            "type": "string",
            "description": "Base64-encoded secret data (encrypted client-side)",
            "example": "bXkgc2VjcmV0IG1lc3NhZ2U="
          },
          "expires_in": {
            "type": "integer",
            "minimum": 1,
            "maximum": 604800,
            "description": "TTL in seconds (max 7 days)",
            "example": 3600
          }
        }
      },
      "CreateSecretResponse": {
        "type": "object",
        "required": ["id"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the created secret",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        }
      },
      "CreateTokenRequest": {
        "type": "object",
        "required": ["ttl_seconds"],
        "properties": {
          "upload_size_limit": {
            "type": "integer",
            "minimum": 1,
            "description": "Optional upload size limit in bytes for the token",
            "example": 1048576
          },
          "ttl_seconds": {
            "type": "integer",
            "minimum": 1,
            "description": "Token time-to-live in seconds",
            "example": 2592000
          }
        }
      },
      "CreateTokenResponse": {
        "type": "object",
        "required": ["token"],
        "properties": {
          "token": {
            "type": "string",
            "description": "The generated authentication token",
            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Authentication token (if required by server configuration)"
      },
      "adminAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Admin authentication token (required for admin operations)"
      }
    }
  },
  "tags": []
}